requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
psycopg2-binary
Flask-JWT-Extended
marshmallow
pytest
pytest-flask